changes: Aed UserDao to solve withdraw servlet and account.jsp the user has to be reseted as current user because of balance issue Added Authentication in accountSummary.jsp Changed customerProfile.jsp Changed Deposit Servlet Added FdRdPortal.jsp Added double balance in loan.jsp Added list of transaction dynamic in transactionHistory.jsp Added getTransactionHistoryList Method in UserDao

+++ src/main/java/com/project/servlets/DepositServlet.java
@@ -30,8 +30,8 @@ public class DepositServlet extends HttpServlet {
                 response.sendRedirect("account.jsp");
 
             } else {
-                int currentBalance = Integer.parseInt(user.getBalance());
-                currentBalance += Integer.parseInt(balance);
+                double currentBalance = Double.parseDouble(user.getBalance());
+                currentBalance += Double.parseDouble(balance);
                 user.setBalance((String.valueOf(currentBalance)));
                 session.update(user);
                 tx.commit();
